+++ src/lib/components/Button/index.ts
@@ -1,0 +1,2 @@
+export { default as Button } from "./Button.svelte";
+export * from "./button";
